Update spec file

+++ go2fed.spec
@@ -1,13 +1,17 @@
-%global commit		9f2c51e270b93d30b133e7f8bc060719432b3a19
+%global provider        github
+%global provider_tld    com
+%global project        	ingvagabund
+%global repo            GolangPackageGenerator
+%global commit		46db1fa5e7c380cd7eddd87c0bb9f1675b8992ea
 %global shortcommit	%(c=%{commit}; echo ${c:0:7})
 
 Name:		go2fed
 Version:	0
-Release:	1%{?dist}
+Release:	4%{?dist}
 Summary:	Tool for development of golang devel packages
 License:	GPLv2+
 URL:		https://github.com/ingvagabund/GolangPackageGenerator
-Source0:	https://github.com/ingvagabund/GolangPackageGenerator/archive/%{commit}/gpg-%{shortcommit}.tar.gz
+Source0:	https://github.com/ingvagabund/GolangPackageGenerator/archive/%{commit}/%{repo}-%{shortcommit}.tar.gz
 
 Requires: python >= 2.7.5, bash, wget, rpmdevtools, rpmlint
 Requires: fedpkg, koji, coreutils, rpm-build, openssh-clients, tar
@@ -20,7 +24,7 @@ dependency discovering, testing (scratch builds), to prepare package review.
 If possible, all in one command.
 
 %prep
-%setup -q -n GolangPackageGenerator-%{commit}
+%setup -q -n %{repo}-%{commit}
 
 %build
 
@@ -35,20 +39,25 @@ cp man/go2fed-help.1 %{buildroot}/usr/share/man/man1/go2fed.1
 mkdir -p %{buildroot}/usr/share/%{name}
 cp *.sh %{buildroot}/usr/share/%{name}/.
 cp *.py %{buildroot}/usr/share/%{name}/.
+cp -r modules %{buildroot}/usr/share/%{name}/.
 # copy config
 mkdir -p %{buildroot}/usr/share/%{name}/config
 cp config/go2fed.conf %{buildroot}/usr/share/%{name}/config/.
 # copy golang list and native imports
-cp golang.list golang.imports %{buildroot}/usr/share/%{name}/.
+cp -r data %{buildroot}/usr/share/%{name}/.
 # copy the tool script
 cp %{name} %{buildroot}/usr/share/%{name}/.
 
 %post
-# make a symlink to go2fed
-ln -s /usr/share/%{name}/%{name} /usr/bin/%{name}
+if [ "$1" -eq 1 ]; then
+	# make a symlink to go2fed
+	ln -s /usr/share/%{name}/%{name} /usr/bin/%{name}
+fi
 
 %preun
-rm /usr/bin/%{name}
+if [ "$1" -eq 0 ]; then
+	rm /usr/bin/%{name}
+fi
 
 %files
 %doc README.md LICENSE
@@ -58,6 +67,15 @@ rm /usr/bin/%{name}
 /usr/share/man/man1/go2fed.1.gz
 
 %changelog
+* Mon Feb 09 2015 jchaloup <jchaloup@redhat.com> - 0-4
+- Bump to upstream 46db1fa5e7c380cd7eddd87c0bb9f1675b8992ea
+
+* Mon Feb 09 2015 jchaloup <jchaloup@redhat.com> - 0-3
+- Bump to upstream f6bbc7c683601e6cbd4bd3c3bceeb616e840511b
+
+* Mon Feb 09 2015 jchaloup <jchaloup@redhat.com> - 0-2
+- Bump to upstream 6a1b91cb818b88e0c40f68b6c8a3bd2ab75b6340
+
 * Fri Oct 24 2014 jchaloup <jchaloup@redhat.com>
 - Initial package
